Progress on live Pizza Count
Pizza Count is nearly live now. It retrieves data from firestore and
in theory will update that with an event listener if that data changes.
That said... It required changing the firestore rules to allow anyone
to read which is not safe for prodcution. Seems like an authentication
problem.

Still to do on Pizza Count:

1. Figure out autentication to allow site to read count value
2. Figure out pub sub
3. Improve "live" => I will make an issue for this but since Youtube
doesn't update the channel stats very often, I think it makes sense to
grab all the videos from YouTube API, parse them to get all the view
counts, and then sum that value and write that to firestore instead.
Need to explore this more to see what it would cost, but seems doable.